+++ firebase-firestore/src/androidTest/java/com/google/firebase/firestore/PipelineTest.java
@@ -144,7 +144,7 @@ public class PipelineTest {
     IntegrationTestUtil.tearDown();
   }
 
-  private final Map<String, Map<String, Object>> bookDocs =
+  static final Map<String, Map<String, Object>> bookDocs =
       mapOfEntries(
           entry(
               "book1",